Update the bit estimation ratio adaptively
Improved the approach proposed in
https://aomedia-review.googlesource.com/c/aom/+/165905
Updated the bit estimation ratio adaptively for each
video instead of using a fixed model.

This BDrate gain is 1.6%(ovr_psnr) and 1.5%(ssim) (the
baseline: accurate_bit_estimate disabled). We found
that this approach works well for videos with background
motion.

Same as prvious CL, this is only enabled for aq_mode=0.

STATS_CHANGED

Change-Id: I30e9af783309297d062f9242788fa3f651fd14d3