Some cleanup related to av1_calc_mb_wiener_var_mt
The code was originally added in
https://aomedia-review.googlesource.com/c/aom/+/166141/3..4.

Pass num_workers as an input parameter to av1_calc_mb_wiener_var_mt() to
avoid calculating the value in both the function and its caller.

In av1_calc_mb_wiener_var_row(), explain why cpi->tile_data may be a
null pointer when the function is called in single-threaded mode. This
removes a TODO(any) comment.

Bug: b/257074966
Change-Id: Ifb45213f2286b99d1024fecf783b8cb23117ce0d