Modify encoder's entropy coding byte-push mechanism
Changed the byte-push mechanism in libaom entropy
encoder, as suggested in a To Do comment. In the
parent version, the register is flushed when at
least 1 byte is available. In the CL, we make the
register 64-bit and flush it when at least 4 bytes
are accumulated in it.

An intermediate 16-bit buffer 'precarry_buf' is
removed and carry-propagation mechanism is changed.

No changes to the decoder logic at this time, it
already has an OPT comment about using uint64_t
for decode-time optimization.

AV1 RTC Encoder:

cpu   Enc time
       red (%)
 5      0.407
 6      0.406
 7      0.612
 8      0.964
 9      1.278
10      1.654

AVIF Encoder:

cpu   Enc time
       red (%)
 6      0.238
 7      0.302
 8      0.664
 9      1.929

This is a bit-exact change.

Change-Id: If23e711145aef1fa193523c98b29c7c82eff76b7