Fix a bug around image pyramid invalidation
In commit 674eaa06, av1_realloc_and_scale_if_required() was changed so
that it invalidated the image pyramid attached to the unscaled frame
if it was going to return that frame.

This change turned out to cause a bug when using frame-parallel
multithreading. The issue was that one thread could call
av1_realloc_and_scale_if_required(), and invalidate the pyramid
attached to a frame buffer, while another thread was still using
that pyramid for global motion estimation.

On further investigation, there is no need for the pyramid to be
invalidated by av1_realloc_and_scale_if_required() in the first place,
so we can just remove that call.

Change-Id: I7454ebe63144104550d38b12c375c43e3f064962